Build `ecad-angular-auth` package before ng serve
The idea being that we would like the `ecad-angular-auth` package to be
built when we start the dev server. This is only half way there
though, as when we are developing, this won't rebuild the package
`ecad-angular-auth` on the fly anyway...

angular-cli is getting a feature in the future to watch and rebuild for
libraries, which will simplify this use case significantly.