Add personal role and add example permissions
The personal role is set as the default role. One single default role is
required to operate. The `default_personal_role` allows for log in to a
personal tenant, and is useful as an intermediary state for a user
before they are added to a tenant, or if they have been removed from a
tenant.

Created an example permission service called `net.example.service` to
illustrate how an operator should add new permissions that are consumed
by downstream services.